docs(uml): add card/shoe/rng class diagram
The intent is for the classes defined here to be applicable to all games targeted for this project, and provide some flexibility for additional games.  These classes are largely model related, and explicitly not view related.  The primary abstraction concern was to not restrict the entire design to using standard playing cards.  However it seems fine that individual games would be restricted to specific card types.