feat(sprite): add playing cards sprite
Asset borrowed from https://xoppa.github.io/blog/a-simple-card-game which was in turn borrowed from https://opengameart.org/content/playing-cards-0.  These assets are licensed as, basically, CC0 (https://creativecommons.org/about/cclicenses/): "Terms & Conditions: The free game assets on this site are released under the Creative Commons Attribution Licence. This basically means that you can download them, use them in your game(s) and modify them as you like. If you use them in your games, we’d appreciate a small mention somewhere in the credits / on the website (and a free copy of the game would be cool!). Basically, the only thing we don’t want you to do with our assets is sell them as assets to other developers, either individually, or as part of a collection."