docs(uml): add notes to banking model uml as I consider overall structure
I'm really considering improving the separation of the various components and the root banking controller.  In particular, it feels like tight coupling that the various components are accessed through the controller (even though that is convenient), and I want to have modal and control portions of the component interfaces so the root implementation can be more abstract.