docs(uml): refactor banking design to separate concerns better
Rather than having a BankController which internally instantiates a wallet, bet settings, and ledger controller I've refactored the design so that the root controller interface makes no reference to these components and instead only exposes the various control operations.  Additionally, currency formatting has been completely abstracted and will be moved to the view side of things, since it is the labels that care about this.  The component interfaces were expanded to allow for better control and more intuitive members.