refactor(console): swap to scriptable objects for console components
The aim is to alleviate the script initialization struggle by not requiring parent controllers to pass references down the chain to child controllers and instead have each child controller have a reference to the nexessary SOs directly.  This goes against one of my desired goals, but I'm not sure that desired goal was a smart one.  Hopefully this experiment works out and I'm not wasting time.