refactor(console): adjust control strip interface so button states can be driven externally
The original API was confused, suggesting some controls were to be controlled externally while others would be managed internally and merely report events.  Adjusted this for consistency so all controls are managed externally (i.e. clicking a button triggers an event and its then up to the owning controller to come back from that event and change the button state.